+++ data/dbQueries.js
@@ -5,4 +5,6 @@ export const dbQueries = {
     "SELECT course.*,attendance.*,sheet.*,groups.ID_GROUP,COUNT(*) AS participants FROM rtc_training AS course INNER JOIN rtc_training_attendance AS attendance ON course.__kp_Course = attendance.training_course_id AND attendance.uploaded_at = '0000-00-00' INNER JOIN rtc_groups AS groups ON attendance.__kf_group = groups.__kp_Group INNER JOIN rtc_attendance_sheets AS sheet ON attendance.uuid = sheet.uuid GROUP BY attendance._kf_training;",
   Q_TMP_GRP_ASSIGN:
     "INSERT INTO tmp_farmer_group_assignment(created_at,farmerid,_kf_farmer,kf_group_old,kf_group_new,group_id_new,_kf_station,assigned_by,uploaded) VALUES",
+  Q_TMP_GRP_ACTVT:
+    "INSERT INTO tmp_group_activate(created_at,_kf_Group,_kf_station,active,username,uploaded) VALUES",
 };
